refactor: use common MetricEntity for Tasklist & Operate
