+++ webapps-schema/src/main/java/io/camunda/webapps/schema/descriptors/tasklist/index/FormIndex.java
@@ -7,6 +7,7 @@
  */
 package io.camunda.webapps.schema.descriptors.tasklist.index;
 
+import io.camunda.webapps.schema.descriptors.IndexDescriptor;
 import io.camunda.webapps.schema.descriptors.backup.Prio4Backup;
 import io.camunda.webapps.schema.descriptors.tasklist.TasklistIndexDescriptor;
 
@@ -19,7 +20,7 @@ public class FormIndex extends TasklistIndexDescriptor implements Prio4Backup {
   public static final String KEY = "key";
   public static final String BPMN_ID = "bpmnId";
   public static final String SCHEMA = "schema";
-  public static final String TENANT_ID = "tenantId";
+  public static final String TENANT_ID = IndexDescriptor.TENANT_ID;
   public static final String VERSION = "version";
   public static final String IS_DELETED = "isDeleted";
