+++ webapps-schema/src/main/java/io/camunda/webapps/schema/entities/MetricEntity.java
@@ -5,7 +5,7 @@
  * Licensed under the Camunda License 1.0. You may not use this file
  * except in compliance with the Camunda License 1.0.
  */
-package io.camunda.webapps.schema.entities.operate;
+package io.camunda.webapps.schema.entities;
 
 import io.camunda.webapps.schema.entities.tasklist.TasklistEntity;
 import java.time.OffsetDateTime;
