+++ zeebe/exporters/camunda-exporter/src/test/java/io/camunda/exporter/handlers/MetricFromProcessInstanceHandlerTest.java
@@ -16,7 +16,7 @@ import static org.mockito.Mockito.verify;
 
 import io.camunda.exporter.store.BatchRequest;
 import io.camunda.webapps.schema.descriptors.operate.index.MetricIndex;
-import io.camunda.webapps.schema.entities.operate.MetricEntity;
+import io.camunda.webapps.schema.entities.MetricEntity;
 import io.camunda.zeebe.protocol.record.Record;
 import io.camunda.zeebe.protocol.record.ValueType;
 import io.camunda.zeebe.protocol.record.intent.ProcessInstanceIntent;
